univer/serializers.py
from rest_framework import serializers
from .models import Course, Grade, Student


class CourseSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Course
        fields = '__all__'


class GradeSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Grade
        fields = '__all__'


class StudentSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Student
        fields = ['id', 'last_name', 'first_name', 'second_name', 'id_number', 'grade_set']
        depth = 1
