requirements.txt
djangorestframework
django-cors-headers
model_mommy